Validated that Client exists by CommerceId and ClientId.

+++ src/main/java/com/equilibrium/webapp/service/PaymentServiceImpl.java
@@ -21,14 +21,22 @@ public class PaymentServiceImpl implements PaymentService {
     private ClientRepository clientRepository;
 
     @Override
-    public Page<Payment> getAllPaymentsByClientId(Long clientId, Pageable pageable) {
-        if(!clientRepository.existsById(clientId))
-            throw new ResourceNotFoundException("Client", "Id", clientId);
+    public Page<Payment> getAllPaymentsByCommerceIdAndClientId(Long commerceId, Long clientId, Pageable pageable) {
+        if(!clientRepository.existsByIdAndCommerceId(clientId, commerceId)){
+            throw new ResourceNotFoundException(
+                    "Client not found with Id " + clientId +
+                            " and CommerceId " + commerceId);
+        }
         return paymentRepository.findByClientId(clientId, pageable);
     }
 
     @Override
-    public Payment getPaymentByIdAndClientId(Long paymentId, Long clientId) {
+    public Payment getPaymentByCommerceIdAndClientIdAndId(Long commerceId, Long clientId, Long paymentId) {
+        if(!clientRepository.existsByIdAndCommerceId(clientId, commerceId)){
+            throw new ResourceNotFoundException(
+                    "Client not found with Id " + clientId +
+                            " and CommerceId " + commerceId);
+        }
         return paymentRepository.findByIdAndClientId(paymentId, clientId)
                 .orElseThrow(() -> new ResourceNotFoundException(
                         "Payment not found with Id " + paymentId +
@@ -36,7 +44,12 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
-    public Payment createPayment(Long clientId, Payment payment) {
+    public Payment createPayment(Long commerceId, Long clientId, Payment payment) {
+        if(!clientRepository.existsByIdAndCommerceId(clientId, commerceId)){
+            throw new ResourceNotFoundException(
+                    "Client not found with Id " + clientId +
+                            " and CommerceId " + commerceId);
+        }
         return clientRepository.findById(clientId).map(client -> {
             payment.setClient(client);
             if (payment.getAmount() > client.getCreditAmount()){
@@ -46,14 +59,17 @@ public class PaymentServiceImpl implements PaymentService {
             client.setCreditAmount(client.getCreditAmount() - payment.getAmount());
             return paymentRepository.save(payment);
         }).orElseThrow(() -> new ResourceNotFoundException(
-                "Client", "Id", clientId));
+                "Client not found with Id " + clientId));
     }
 
     @Override
-    public Payment updatePayment(Long clientId, Long paymentId, Payment request) {
-        if(!clientRepository.existsById(clientId))
-            throw new ResourceNotFoundException("Client", "Id", clientId);
-        return paymentRepository.findById(paymentId).map(payment -> {
+    public Payment updatePayment(Long commerceId, Long clientId, Long paymentId, Payment request) {
+        if(!clientRepository.existsByIdAndCommerceId(clientId, commerceId)){
+            throw new ResourceNotFoundException(
+                    "Client not found with Id " + clientId +
+                            " and CommerceId " + commerceId);
+        }
+        return paymentRepository.findByIdAndClientId(paymentId, clientId).map(payment -> {
             payment.setAmount(request.getAmount());
             payment.setDescription(request.getDescription());
             return paymentRepository.save(payment);
@@ -62,7 +78,12 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
-    public ResponseEntity<?> deletePayment(Long paymentId, Long clientId) {
+    public ResponseEntity<?> deletePayment(Long commerceId, Long clientId, Long paymentId) {
+        if(!clientRepository.existsByIdAndCommerceId(clientId, commerceId)){
+            throw new ResourceNotFoundException(
+                    "Client not found with Id " + clientId +
+                            " and CommerceId " + commerceId);
+        }
         return paymentRepository.findByIdAndClientId(paymentId, clientId).map(payment -> {
             paymentRepository.delete(payment);
             return ResponseEntity.ok().build();
